requirements.txt
django
djangorestframework
markdown
django-filter
django-guardian
drfdocs
django-rest-framework-social-oauth2